exercises4.15-4.21

+++ part4/blog-list/controllers/blogs.js
@@ -1,21 +1,50 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  const blog = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-  const result = await blog.save()
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  const newBlog = new Blog({
+    title: blog.title,
+    author: blog. author,
+    url: blog.url,
+    likes: blog.likes,
+    user: user._id
+  })
+
+  const result = await newBlog.save()
+  user.blogs = user.blogs.concat(result.id)
+  await user.save()
   response.status(201).json(result)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const { id: userId } = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+
+  if(userId === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'token or user are invalid' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
